LPS-167572 SF, move logic to DirectoryWatcher
